agent/requirements.txt
agent-framework-core==1.0.0b260107
azure-ai-agentserver-core==1.0.0b10
azure-ai-agentserver-agentframework==1.0.0b10
python-dotenv>=1.0.1
httpx>=0.27.0
